Moved LanguageContext mock values for testing to a separate component
