+++ frontend/src/__tests__/components/LanguagePicker.test.jsx
@@ -1,40 +1,23 @@
-import { vi, test, expect } from "vitest";
+import { test, expect } from "vitest";
 import { render, screen, fireEvent } from "@testing-library/react";
 import { LanguagePicker } from "../../components/LanguagePicker";
 import { AuthContext } from "../../context/AuthContext.jsx";
 import { LanguageContext } from "../../context/LanguageContext.jsx";
 import { BrowserRouter as Router } from "react-router-dom";
-import { I18nextProvider } from "react-i18next";
-import i18n from "../../i18n-test.js";
-
-const languages = [
-  { id: 1, code: "en", name: "English" },
-  { id: 2, code: "jp", name: "日本語" },
-  { id: 3, code: "fi", name: "Suomi" },
-];
-
-const languageContextValue = {
+import {
   languages,
-  selectedLanguage: "en",
-  handleLanguageChange: (newLanguage) => {
-    languageContextValue.selectedLanguage = newLanguage;
-    i18n.changeLanguage(newLanguage);
-  },
-  updateLanguage: vi.fn(),
-  t: i18n.t,
-};
+  languageContextValue,
+} from "../../utils/languageTestSetup.js";
 
 test("renders LanguagePicker component", () => {
   render(
-    <I18nextProvider i18n={i18n}>
-      <AuthContext.Provider value={{ isLoggedIn: true }}>
-        <LanguageContext.Provider value={languageContextValue}>
-          <Router>
-            <LanguagePicker />
-          </Router>
-        </LanguageContext.Provider>
-      </AuthContext.Provider>
-    </I18nextProvider>
+    <AuthContext.Provider value={{ isLoggedIn: true }}>
+      <LanguageContext.Provider value={languageContextValue}>
+        <Router>
+          <LanguagePicker />
+        </Router>
+      </LanguageContext.Provider>
+    </AuthContext.Provider>
   );
 
   const languagePickerElement = screen.getByTestId("language-picker");
@@ -46,15 +29,13 @@ test("renders LanguagePicker component", () => {
 
 test("renders all languages", () => {
   render(
-    <I18nextProvider i18n={i18n}>
-      <AuthContext.Provider value={{ isLoggedIn: true }}>
-        <LanguageContext.Provider value={languageContextValue}>
-          <Router>
-            <LanguagePicker />
-          </Router>
-        </LanguageContext.Provider>
-      </AuthContext.Provider>
-    </I18nextProvider>
+    <AuthContext.Provider value={{ isLoggedIn: true }}>
+      <LanguageContext.Provider value={languageContextValue}>
+        <Router>
+          <LanguagePicker />
+        </Router>
+      </LanguageContext.Provider>
+    </AuthContext.Provider>
   );
 
   languages.forEach((language) => {
@@ -65,15 +46,13 @@ test("renders all languages", () => {
 
 test("changes language to japanese when a new option is selected", async () => {
   const { rerender } = render(
-    <I18nextProvider i18n={i18n}>
-      <AuthContext.Provider value={{ isLoggedIn: true }}>
-        <LanguageContext.Provider value={languageContextValue}>
-          <Router>
-            <LanguagePicker />
-          </Router>
-        </LanguageContext.Provider>
-      </AuthContext.Provider>
-    </I18nextProvider>
+    <AuthContext.Provider value={{ isLoggedIn: true }}>
+      <LanguageContext.Provider value={languageContextValue}>
+        <Router>
+          <LanguagePicker />
+        </Router>
+      </LanguageContext.Provider>
+    </AuthContext.Provider>
   );
 
   const languagePickerElement = screen.getByTestId("language-picker");
@@ -82,15 +61,13 @@ test("changes language to japanese when a new option is selected", async () => {
   languageContextValue.selectedLanguage = "jp";
 
   rerender(
-    <I18nextProvider i18n={i18n}>
-      <AuthContext.Provider value={{ isLoggedIn: true }}>
-        <LanguageContext.Provider value={languageContextValue}>
-          <Router>
-            <LanguagePicker />
-          </Router>
-        </LanguageContext.Provider>
-      </AuthContext.Provider>
-    </I18nextProvider>
+    <AuthContext.Provider value={{ isLoggedIn: true }}>
+      <LanguageContext.Provider value={languageContextValue}>
+        <Router>
+          <LanguagePicker />
+        </Router>
+      </LanguageContext.Provider>
+    </AuthContext.Provider>
   );
 
   const jpTitleElement = screen.getByText("言語を選択:");
@@ -99,15 +76,13 @@ test("changes language to japanese when a new option is selected", async () => {
 
 test("changes language to finnish when a new option is selected", async () => {
   const { rerender } = render(
-    <I18nextProvider i18n={i18n}>
-      <AuthContext.Provider value={{ isLoggedIn: true }}>
-        <LanguageContext.Provider value={languageContextValue}>
-          <Router>
-            <LanguagePicker />
-          </Router>
-        </LanguageContext.Provider>
-      </AuthContext.Provider>
-    </I18nextProvider>
+    <AuthContext.Provider value={{ isLoggedIn: true }}>
+      <LanguageContext.Provider value={languageContextValue}>
+        <Router>
+          <LanguagePicker />
+        </Router>
+      </LanguageContext.Provider>
+    </AuthContext.Provider>
   );
 
   const languagePickerElement = screen.getByTestId("language-picker");
@@ -116,15 +91,13 @@ test("changes language to finnish when a new option is selected", async () => {
   languageContextValue.selectedLanguage = "fi";
 
   rerender(
-    <I18nextProvider i18n={i18n}>
-      <AuthContext.Provider value={{ isLoggedIn: true }}>
-        <LanguageContext.Provider value={languageContextValue}>
-          <Router>
-            <LanguagePicker />
-          </Router>
-        </LanguageContext.Provider>
-      </AuthContext.Provider>
-    </I18nextProvider>
+    <AuthContext.Provider value={{ isLoggedIn: true }}>
+      <LanguageContext.Provider value={languageContextValue}>
+        <Router>
+          <LanguagePicker />
+        </Router>
+      </LanguageContext.Provider>
+    </AuthContext.Provider>
   );
 
   const fiTitleElement = screen.getByText("Valitse kieli:");
