+++ frontend/src/utils/languageTestSetup.js
@@ -1,0 +1,21 @@
+import i18nInstance from "../i18n-test";
+import { vi } from "vitest";
+
+export const i18n = i18nInstance;
+
+export const languages = [
+  { id: 1, code: "en", name: "English" },
+  { id: 2, code: "jp", name: "日本語" },
+  { id: 3, code: "fi", name: "Suomi" },
+];
+
+export const languageContextValue = {
+  languages,
+  selectedLanguage: "en",
+  handleLanguageChange: (newLanguage) => {
+    languageContextValue.selectedLanguage = newLanguage;
+    i18n.changeLanguage(newLanguage);
+  },
+  updateLanguage: vi.fn(),
+  t: i18n.t,
+};
